+++ app/routes/RoomRoutes.js
@@ -1,9 +1,21 @@
 ﻿module.exports = function (app) {
 
     var RoomService = require('./../services/RoomService');
+    var PermissionService = require('./../services/PermissionService');
+    var ValidationResult = require('./../models/ValidationResultStructure');
+    var PermissionModule = require('./../models/StatModel').PermissionModule;
+    var PermissionType = require('./../models/StatModel').PermissionType;
 
     // save room
     app.post('/api/room', function (req, res) {
+        // Check if user has permissions
+        if (!PermissionService.check(true, req.session, PermissionModule.ADMINISTRATION_ROOM, PermissionType.WRITE)) {
+            var validation = new ValidationResult({});
+            validation.addError('Uživatel nemá oprávnění ke zvolené operaci');
+            res.json(validation);
+            return;
+        }
+
         RoomService.save(req.body, function (validation) {
             res.json(validation);
         });
@@ -18,20 +30,29 @@
 
     // get all rooms
     app.get('/api/room', function (req, res) {
-        RoomService.getList(function (validation) {
+        // Check if user has permissions
+        if (!PermissionService.check(true, req.session, PermissionModule.ADMINISTRATION_ROOM, PermissionType.READ)) {
+            var validation = new ValidationResult({});
+            validation.addError('Uživatel nemá oprávnění ke zvolené operaci');
             res.json(validation);
-        });
-    });
+            return;
+        }
 
-    // get available rooms
-    app.post('/api/room/available', function (req, res) {
-        RoomService.getAvailableRooms(req.body, function (validation) {
+        RoomService.getList(function (validation) {
             res.json(validation);
         });
     });
 
     // get room
     app.get('/api/room/:room_id', function (req, res) {
+        // Check if user has permissions
+        if (!PermissionService.check(true, req.session, PermissionModule.ADMINISTRATION_ROOM, PermissionType.READ)) {
+            var validation = new ValidationResult({});
+            validation.addError('Uživatel nemá oprávnění ke zvolené operaci');
+            res.json(validation);
+            return;
+        }
+
         RoomService.get({
             _id: req.params.room_id
         }, function (validation) {
@@ -41,6 +62,14 @@
 
     // delete room
     app.delete('/api/room/:room_id', function (req, res) {
+        // Check if user has permissions
+        if (!PermissionService.check(true, req.session, PermissionModule.ADMINISTRATION_ROOM, PermissionType.WRITE)) {
+            var validation = new ValidationResult({});
+            validation.addError('Uživatel nemá oprávnění ke zvolené operaci');
+            res.json(validation);
+            return;
+        }
+
         RoomService.remove({
             _id: req.params.room_id
         }, function (validation) {
